Add live debugger sender FFI
