+++ live-debugger-ffi/src/lib.rs
@@ -4,4 +4,5 @@
 pub mod data;
 pub mod evaluator;
 pub mod parse;
+pub mod send_data;
 pub mod sender;
